pyproject.toml
[tool.poetry]
name = "hng3"
version = "0.1.0"
description = ""
authors = ["kalio007 <kalioprince1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
rabbitmq = "^0.2.0"
pika = "^1.3.2"
aiofiles = "^24.1.0"
celery = {extras = ["redis"], version = "^5.4.0"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
